requirements.txt
django==5.2.3
django-encrypted-model-fields==0.6.5
django-picklefield==3.3
django-q==1.3.9

Pillow==9.4.0
cryptography==45.0.4
python-escpos==3.1
python-dotenv==1.0.1
python-decouple==3.8

google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.182.0
google-auth-httplib2==0.2.0

pymupdf==1.26.3
bcrypt==4.1.3

psycopg2==2.9.5
sqlite-fts4==1.0.3
sqlite-utils==3.38
sqlparse==0.5.3

openpyxl==3.1.5

celery==5.6.0
redis>=3.5.3,<4.0.0

django-minify-html==1.14.0